refactor: concluir migração de Double para BigDecimal em fluxo financeiro

+++ app/src/main/java/com/migueldk17/breeze/ui/features/adicionarconta/ui/layouts/Passo4.kt
@@ -36,6 +36,7 @@ import com.migueldk17.breeze.ui.features.adicionarconta.ui.components.Personaliz
 import com.migueldk17.breeze.ui.features.adicionarconta.ui.components.ResponsiveDateParcelaSection
 import com.migueldk17.breeze.ui.features.adicionarconta.viewmodels.AdicionarContaViewModel
 import com.migueldk17.breeze.ui.features.paginainicial.ui.components.BreezeDatePicker
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @SuppressLint("UnusedBoxWithConstraintsScope")
@@ -168,7 +169,7 @@ fun Passo4(
                     .padding(vertical = 74.dp),
                 text = "Avançar",
                 onClick = {
-                    viewModel.guardaValorConta(valorConta.toDouble())
+                    viewModel.guardaValorConta(valorConta.toBigDecimal())
                     if (textJuros != "") viewModel.guardaPorcentagemJuros(textJuros) //Guarda a porcentagem de juros
                     viewModel.guardaDataConta(selectedDate) //Guarda a data da conta
                     if (textParcelas.isEmpty()) viewModel.guardaQtdParcelas(selectedCategory) else viewModel.guardaQtdParcelas(textParcelas) //Guarda a quantidade de parcelas
